fix: change how scenario uses input parameters
Now all the parameters are added along with the executable.
If you want the executable to be run few times with different parameters
you need to add a line for each execution.
This way you may execute programs with different input, more or less
arguments. This is first step to add unreal tests which are parametrized
with ini file