feat: allow python server side in test runner
python tests (and unreal) are run differently than cpp and qt based tests
they are not just an application to start, but require a command to run
test runner is modified in a way it allows now run server as shell command
The python server does not have an app that checking messages to know
when the client finished. It is kill when client is over.