chores: clean up structure
the network adapters files are very imported files for test - this is
not a heloer file, it should be more visible.
Most of helepr files are used for many object tests and were move there
aprat from olink, where it is still used for single object test, but
probably should not.
This is just a one step of clean up.
The latency helpers (common for all tests) should be separate target
The olink test may be improved.
The TestData objects are very common between technologies, probably
could be taken out.